requirements.txt
# Core PyTorch ecosystem
torch>=2.5.0
torchvision>=0.20.0

# Core ML/AI libraries
# Note: numpy<2.0.0 to avoid compatibility issues with detectron2/densepose
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
scikit-image>=0.20.0
# Use opencv-python (not headless) and pin to version compatible with numpy<2.0
opencv-python>=4.8.0,<4.12.0
pillow>=10.0.0
onnxruntime>=1.15.0

# Optional ML libraries (commented out if installation fails)
# These are used by detectron2/densepose but may not be compatible with all torch versions
# diffusers>=0.20.2
# transformers>=4.30.0
# accelerate>=0.20.0
# einops>=0.6.0
# xformers>=0.0.20  # Not compatible with torch 2.7.1, install separately if needed

# Optional utilities (commented out if installation fails)
# basicsr>=1.4.2  # Used by openpose, but optional (has fallback)
# av>=10.0.0  # Video processing, optional
# gradio==3.44.4  # UI framework, optional
# matplotlib>=3.7.0  # Plotting, optional
# tqdm>=4.65.0  # Progress bars, optional

# Core dependencies for detectron2 and densepose
cloudpickle>=2.2.0
fvcore>=0.1.5
hydra-core>=1.1.0,<1.4.0
pycocotools>=2.0.6

# Additional dependencies for detectron2 and densepose
setuptools>=65.0.0
wheel>=0.40.0
pyyaml>=6.0
yacs>=0.1.8
tabulate>=0.9.0
termcolor>=2.3.0
iopath>=0.1.9
portalocker>=2.7.0
pydot>=1.4.2

# Note: detectron2 and densepose will be installed from local source
# They are located in preprocess/detectron2/

# Kafka messaging (using message_broker)
# Note: Install message_broker from ../kafka_component first:
#   cd ../kafka_component && pip install -e .
confluent-kafka>=2.1.0

# MinIO/S3 storage
minio>=7.2.0

# PostgreSQL async driver
asyncpg>=0.29.0

# HTTP client for image downloads
aiohttp>=3.9.0

# Retry and backoff utilities
tenacity>=8.2.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
